requirements.txt
faker
pytest
allure-pytest
requests